+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: `OpenLockIconNext`: New component.
+- Feature: `DashboardLayout`: Add the following subcomponents to `DashboardLayout.Flow`:
+  - `Title`
+  - `TitleAside`
+- Fix: `LightbulbIllustration`: Fix colors.
+- Fix: `TextCopy`: Better className propagation.
 
 ## [9.5.0] - 2022-04-01