+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Change/Deprecation: `Typography`: 
   - Replace `tiny, regular, default, big` by `small, medium, large` prop size values.
+- Feature: `OpenLockIconNext`: New component.
+- Feature: `DashboardLayout`: Add the following subcomponents to `DashboardLayout.Flow`:
+  - `Title`
+  - `TitleAside`
+- Fix: `LightbulbIllustration`: Fix colors.
+- Fix: `TextCopy`: Better className propagation.
 
 ## [9.5.0] - 2022-04-01